move functions to their appropriate locations
